+++ lib/widget_utils.dart
@@ -33,3 +33,4 @@ class PageLoader extends StatelessWidget {
     return loader;
   }
 }
+